+++ README.md
@@ -1,0 +1,33 @@
+# Diplom_1
+
+
+Задание 1: юнит-тесты
+Представь: нужно протестировать программу, которая помогает заказать бургер в Stellar Burgers. 
+Тебе предстоит покрыть её юнит-тестами.
+
+Здесь пригодятся моки, стабы и параметризация: где именно их использовать, реши самостоятельно.
+
+Что нужно сделать
+Склонируй репозиторий с заготовкой кода. (https://github.com/yandex-praktikum/QA-java-diplom-1)
+Подключи библиотеки: Jacoco, Mockito, JUnit 4.
+
+Покрой тестами классы Bun, Burger, Ingredient, IngredientType. 
+Используй моки, стабы и параметризацию там, где нужно.
+
+Процент покрытия должен быть не ниже 70%.
+
+Как выполнить и сдать работу
+Склонируй репозиторий с заготовкой кода, допиши программу в IDEA и запушь код на GitHub.
+
+
+
+МОИ ШАГИ:
+
+Создал отдельный репозиторий для юнит-тестов.
+Создал Maven-проект.
+Подключил Jacoco, Mockito, JUnit 4.
+
+Проект:
+1. Покрыл тестами классы Bun, Burger, Ingredient, IngredientType. 
+2. Написал параметризацию в классе BunParameterTest
+3. Сделан отчёт с помощью Jacoco.
