+++ src/main/java/praktikum/praktikum/Ingredient.java
@@ -1,0 +1,32 @@
+package praktikum;
+
+/**
+ * Модель ингредиента.
+ * Ингредиент: начинка или соус.
+ * У ингредиента есть тип (начинка или соус), название и цена.
+ */
+public class Ingredient {
+
+    public IngredientType type;
+    public String name;
+    public float price;
+
+    public Ingredient(IngredientType type, String name, float price) {
+        this.type = type;
+        this.name = name;
+        this.price = price;
+    }
+
+    public float getPrice() {
+        return price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public IngredientType getType() {
+        return type;
+    }
+
+}
