+++ src/main/java/praktikum/praktikum/IngredientType.java
@@ -1,0 +1,11 @@
+package praktikum;
+
+/**
+ * Перечисление с типами ингредиентов.
+ * SAUCE – соус
+ * FILLING – начинка
+ */
+public enum IngredientType {
+    SAUCE,
+    FILLING
+}
