+++ src/test/java/BurgerTest.java
@@ -1,0 +1,64 @@
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.MockitoJUnitRunner;
+import praktikum.*;
+import java.util.List;
+import static org.junit.Assert.assertEquals;
+
+
+@RunWith(MockitoJUnitRunner.class) // запустили этот класс с Mockito
+public class BurgerTest {
+
+    //Чтобы создать мок, используй аннотацию @Mock — перед Bun bun, Ingredient ingredient и Ingredient ingredient2
+    @Mock
+    Bun bun;
+    @Mock
+    Ingredient ingredient;
+    @Mock
+    Ingredient ingredientSecond;
+    @Mock
+    Database database;
+
+
+    private final Burger burger = new Burger();
+    private final List<Bun> buns = List.of(new Bun("grey bun", 100.50F));
+    private final String bunName = "grey bun";
+
+
+    @Before
+    public void setDefaultBun() { // установить булочку по умолчанию
+        Mockito.when(database.availableBuns()).thenReturn(buns);
+    }
+
+    @Test
+    public void checkSetBuns() { // проверить набор булок
+        burger.setBuns(database.availableBuns().get(0));
+        assertEquals(bunName, burger.bun.getName());
+    }
+
+    @Test
+    public void getPriceCalculateCorrectPrice() { // получить цену. Рассчитать правильную цену
+        Mockito.when(ingredient.getPrice()).thenReturn(125.5F);
+        Mockito.when(ingredientSecond.getPrice()).thenReturn(250F);
+        burger.setBuns(database.availableBuns().get(0));
+        burger.addIngredient(ingredient);
+        burger.addIngredient(ingredientSecond);
+        float expectedBurgerPrice = 576.5F;
+        assertEquals("Неправильная цена бургера с двумя добавленными ингредиентами", expectedBurgerPrice, burger.getPrice(), 0);
+    }
+
+    @Test
+    public void getReceipt() { // получить рецепт
+        burger.addIngredient(new Ingredient(IngredientType.SAUCE, "sour cream", 200));
+        burger.addIngredient(new Ingredient(IngredientType.FILLING, "cutlet", 100));
+        burger.setBuns(bun);
+        String expectedReceipt = burger.getReceipt();
+        assertEquals(expectedReceipt, burger.getReceipt());
+    }
+
+}
+
+
